requirements.txt
python-dotenv
pydantic
langchain
langchain_google_genai
jinja2
langgraph
IPython

